Use gradle for java import tool
-Change structure of project to match gradle structure and add a package
-Remove .classpath and .project -> they can be generated with gradle
eclipse
-Add build.gradle script

With this we no longer have to manually download the required libs and
run the adapt_classpath script it is all done when calling gradle build
