--- JavaImportTool/.classpath
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.6"/>
-	<classpathentry kind="lib" path="/opt/eclipse/plugins/org.eclipse.jdt.core_3.11.2.v20160128-0629.jar"/>
-	<classpathentry kind="lib" path="/opt/eclipse/plugins/org.eclipse.core.runtime_3.11.1.v20150903-1804.jar"/>
-	<classpathentry kind="lib" path="/opt/eclipse/plugins/org.eclipse.core.jobs_3.7.0.v20150330-2103.jar"/>
-	<classpathentry kind="lib" path="/opt/eclipse/plugins/org.eclipse.core.contenttype_3.5.0.v20150421-2214.jar"/>
-	<classpathentry kind="lib" path="/opt/eclipse/plugins/org.eclipse.core.resources_3.10.1.v20150725-1910.jar"/>
-	<classpathentry kind="lib" path="/opt/eclipse/plugins/org.eclipse.equinox.common_3.7.0.v20150402-1709.jar"/>
-	<classpathentry kind="lib" path="/opt/eclipse/plugins/org.eclipse.equinox.preferences_3.5.300.v20150408-1437.jar"/>
-	<classpathentry kind="lib" path="/opt/eclipse/plugins/org.eclipse.osgi_3.10.102.v20160118-1700.jar"/>
-	<classpathentry kind="lib" path="/opt/eclipse/plugins/org.eclipse.text_3.5.400.v20150505-1044.jar"/>
-	<classpathentry kind="lib" path="lib/commons-lang3-3.3.2.jar"/>
-	<classpathentry kind="lib" path="lib/commons-io-2.4.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
