+++ JavaImportTool/src/main/java/envision/java/importtool/MethodMetrics.java
@@ -1,4 +1,4 @@
-package javaImportTool;
+package envision.java.importtool;
 
 import java.util.ArrayList;
 import java.util.List;
