+++ JavaImportTool/src/main/java/envision/java/importtool/Node.java
@@ -23,7 +23,7 @@
  ** OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  **
  **********************************************************************************************************************/
-package javaImportTool;
+package envision.java.importtool;
 
 import java.io.File;
 import java.io.FileNotFoundException;
