+++ JavaImportTool/src/main/java/envision/java/importtool/SizeEstimator.java
@@ -1,4 +1,4 @@
-package javaImportTool;
+package envision.java.importtool;
 
 import java.util.LinkedList;
 import java.util.List;
