+++ JavaImportTool/src/main/java/envision/java/importtool/UnknownFeatureException.java
@@ -23,18 +23,17 @@
  ** OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  **
  **********************************************************************************************************************/
-package javaImportTool;
+package envision.java.importtool;
 
-public class ConversionException extends Exception
-{
-	private static final long serialVersionUID = 8737019016361917805L;
+public class UnknownFeatureException extends ConversionException {
 
-	public ConversionException(String message) {
+	private static final long serialVersionUID = -1487628539179326028L;
+
+	public UnknownFeatureException(String message) {
 		super(message);
 	}
 
-	public ConversionException(String message, Throwable throwable) {
+	public UnknownFeatureException(String message, Throwable throwable) {
 		super(message, throwable);
 	}
 }
-
